match.py: --tc-seconds accepts floats
A fractional base clock (e.g. 2.5s hyper-TC) was silently impossible from
the CLI -- int('2.5') raised -- while --tc-increment already took floats.
Config default stays int; all consumers (init_ms, tc_label) already
handle floats. Verified with a 2.5+0.05 micro-match.

+++ match.py
@@ -839,7 +839,10 @@ def main():
             MODE = argv[i + 1]
             i += 2
         elif argv[i] == "--tc-seconds" and i + 1 < len(argv):
-            TC_SECONDS = int(argv[i + 1])
+            # float, not int: a fractional base clock (e.g. 7.5s hyper-TC)
+            # was silently impossible from the CLI while --tc-increment
+            # already took floats.
+            TC_SECONDS = float(argv[i + 1])
             i += 2
         elif argv[i] == "--tc-increment" and i + 1 < len(argv):
             TC_INCREMENT = float(argv[i + 1])
